+++ qanything_kernel/qanything_server/handler.py
@@ -29,6 +29,7 @@ import base64
 import jwt
 from datetime import timedelta
 import bcrypt
+from qanything_kernel.qanything_server.auth import auth_required
 
 __all__ = ["new_knowledge_base", "upload_files", "list_kbs", "list_docs", "delete_knowledge_base", "delete_docs",
            "rename_knowledge_base", "get_total_status", "clean_files_by_status", "upload_weblink", "local_doc_chat",
@@ -43,10 +44,6 @@ INVALID_USER_ID = f"fail, Invalid user_id: . user_id 必须只含有字母，数
 GATEWAY_IP = os.getenv("GATEWAY_IP", "localhost")
 debug_logger.info(f"GATEWAY_IP: {GATEWAY_IP}")
 
-# JWT配置
-JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY", "your-secret-key-here")  # 生产环境中应使用环境变量
-JWT_ALGORITHM = "HS256"
-JWT_ACCESS_TOKEN_EXPIRES = timedelta(hours=24)  # 令牌有效期
 
 # 异步包装器，用于在后台执行带有参数的同步函数
 async def run_in_background(func, *args):
@@ -69,954 +66,6 @@ def sync_function_with_args(arg1, arg2):
     time.sleep(5)
     print(f"同步函数执行完毕，参数值：arg1={arg1}, arg2={arg2}")
     
-# 创建一个装饰器用于接口权限控制
-def auth_required(permission_level="read", check_kb_access=False):
-    """
-    权限控制装饰器
-    :param permission_level: 所需权限级别（read/write/admin）
-    :param check_kb_access: 是否检查知识库访问权限（针对特定知识库的操作）
-    """
-    def decorator(f):
-        @functools.wraps(f)
-        async def decorated_function(req, *args, **kwargs):
-            # 获取用户ID和令牌
-            user_id = safe_get(req, 'user_id')
-            token = req.headers.get('Authorization', '').replace('Bearer ', '')
-            
-            debug_logger.info(f"权限验证开始 - 用户: {user_id}, 所需权限: {permission_level}, 检查知识库权限: {check_kb_access}")
-            
-            if not user_id:
-                debug_logger.error("未提供用户ID")
-                return response.json({"code": 401, "msg": "未提供用户ID"})
-                
-            if not token:
-                debug_logger.error(f"用户 {user_id} 未提供认证令牌")
-                return response.json({"code": 401, "msg": "未提供认证令牌"})
-                
-            # 验证令牌
-            try:
-                # 解码JWT令牌
-                payload = jwt.decode(token, JWT_SECRET_KEY, algorithms=[JWT_ALGORITHM])
-                
-                # 验证令牌中的用户ID与请求中的用户ID是否匹配
-                token_user_id = payload.get("user_id")
-                if token_user_id != user_id:
-                    debug_logger.error(f"令牌与用户ID不匹配 - 令牌用户: {token_user_id}, 请求用户: {user_id}")
-                    return response.json({"code": 401, "msg": "令牌与用户ID不匹配"})
-                
-                # 检查令牌是否过期
-                exp = payload.get("exp")
-                if not exp or datetime.utcfromtimestamp(exp) < datetime.utcnow():
-                    debug_logger.error(f"用户 {user_id} 的令牌已过期")
-                    return response.json({"code": 401, "msg": "令牌已过期"})
-                
-                # 检查用户角色和权限
-                user_role = payload.get("role") or get_user_role(user_id, req)
-                debug_logger.info(f"用户角色验证 - 用户: {user_id}, 角色: {user_role}, 所需权限: {permission_level}")
-                
-                # 根据角色和请求的权限级别判断是否有权限
-                if not has_permission(user_role, permission_level):
-                    debug_logger.error(f"权限不足 - 用户: {user_id}, 角色: {user_role}, 所需权限: {permission_level}")
-                    return response.json({"code": 403, "msg": f"没有{permission_level}权限执行此操作"})
-                
-                # 如果需要检查知识库访问权限
-                if check_kb_access:
-                    kb_id = safe_get(req, 'kb_id')
-                    if kb_id:
-                        local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-                        # 检查用户是否有权限访问该知识库
-                        if not local_doc_qa.milvus_summary.check_kb_access(user_id, kb_id, permission_level):
-                            debug_logger.error(f"知识库访问权限不足 - 用户: {user_id}, 知识库: {kb_id}, 所需权限: {permission_level}")
-                            return response.json({"code": 403, "msg": f"没有权限{permission_level}访问知识库 {kb_id}"})
-                    
-                    # 如果是批量操作多个知识库
-                    kb_ids = safe_get(req, 'kb_ids')
-                    if kb_ids and isinstance(kb_ids, list):
-                        local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-                        for kb_id in kb_ids:
-                            if not local_doc_qa.milvus_summary.check_kb_access(user_id, kb_id, permission_level):
-                                debug_logger.error(f"知识库访问权限不足 - 用户: {user_id}, 知识库: {kb_id}, 所需权限: {permission_level}")
-                                return response.json({"code": 403, "msg": f"没有权限{permission_level}访问知识库 {kb_id}"})
-                
-                # 将用户信息添加到请求上下文中，方便后续使用
-                req.ctx.user = {
-                    "user_id": user_id,
-                    "role": user_role
-                }
-                
-                debug_logger.info(f"权限验证通过 - 用户: {user_id}, 角色: {user_role}")
-                return await f(req, *args, **kwargs)
-            except jwt.ExpiredSignatureError:
-                debug_logger.error(f"用户 {user_id} 的令牌已过期")
-                return response.json({"code": 401, "msg": "令牌已过期"})
-            except jwt.InvalidTokenError:
-                debug_logger.error(f"用户 {user_id} 提供了无效的令牌")
-                return response.json({"code": 401, "msg": "无效的令牌"})
-            except Exception as e:
-                debug_logger.error(f"认证失败 - 用户: {user_id}, 错误: {str(e)}")
-                return response.json({"code": 401, "msg": f"认证失败: {str(e)}"})
-        return decorated_function
-    return decorator
-
-
-@get_time_async
-@auth_required("write")
-async def new_knowledge_base(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    # user_info = safe_get(req, 'user_info', "1234")
-    # passed, msg = check_user_id_and_user_info(user_id, user_info)
-    # if not passed:
-        # return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("new_knowledge_base %s", user_id)
-    kb_name = safe_get(req, 'kb_name')
-    debug_logger.info("kb_name: %s", kb_name)
-    default_kb_id = 'KB' + uuid.uuid4().hex
-    kb_id = safe_get(req, 'kb_id', default_kb_id)
-    kb_id = correct_kb_id(kb_id)
-
-    is_quick = safe_get(req, 'quick', False)
-    if is_quick:
-        kb_id += "_QUICK"
-
-    if kb_id[:2] != 'KB':
-        return sanic_json({"code": 2001, "msg": "fail, kb_id must start with 'KB'"})
-    not_exist_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, [kb_id])
-    if not not_exist_kb_ids:
-        return sanic_json({"code": 2001, "msg": "fail, knowledge Base {} already exist".format(kb_id)})
-
-    # local_doc_qa.create_milvus_collection(user_id, kb_id, kb_name)
-    local_doc_qa.milvus_summary.new_milvus_base(kb_id, user_id, kb_name)
-    now = datetime.now()
-    timestamp = now.strftime("%Y%m%d%H%M")
-    return sanic_json({"code": 200, "msg": "success create knowledge base {}".format(kb_id),
-                       "data": {"kb_id": kb_id, "kb_name": kb_name, "timestamp": timestamp}})
-
-
-@get_time_async
-async def upload_weblink(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("upload_weblink %s", user_id)
-    debug_logger.info("user_info %s", user_info)
-    kb_id = safe_get(req, 'kb_id')
-    kb_id = correct_kb_id(kb_id)
-    not_exist_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, [kb_id])
-    if not_exist_kb_ids:
-        msg = "invalid kb_id: {}, please check...".format(not_exist_kb_ids)
-        return sanic_json({"code": 2001, "msg": msg, "data": [{}]})
-
-    url = safe_get(req, 'url')
-    if url:
-        urls = [url]
-        # 如果URL以/结尾，先去除这个/
-        if url.endswith('/'):
-            url = url[:-1]
-        titles = [safe_get(req, 'title', url.split('/')[-1]) + '.web']
-    else:
-        urls = safe_get(req, 'urls')
-        titles = safe_get(req, 'titles')
-        if len(urls) != len(titles):
-            return sanic_json({"code": 2003, "msg": "fail, urls and titles length not equal"})
-
-    for url in urls:
-        # url 需要以http开头
-        if not url.startswith('http'):
-            return sanic_json({"code": 2001, "msg": "fail, url must start with 'http'"})
-        # url 长度不能超过2048
-        if len(url) > 2048:
-            return sanic_json({"code": 2002, "msg": f"fail, url too long, max length is 2048."})
-
-    file_names = []
-    for title in titles:
-        debug_logger.info('ori name: %s', title)
-        file_name = re.sub(r'[\uFF01-\uFF5E\u3000-\u303F]', '', title)
-        debug_logger.info('cleaned name: %s', file_name)
-        file_name = truncate_filename(file_name, max_length=110)
-        file_names.append(file_name)
-
-    mode = safe_get(req, 'mode', default='soft')  # soft代表不上传同名文件，strong表示强制上传同名文件
-    debug_logger.info("mode: %s", mode)
-    chunk_size = safe_get(req, 'chunk_size', default=DEFAULT_PARENT_CHUNK_SIZE)
-    debug_logger.info("chunk_size: %s", chunk_size)
-
-    exist_file_names = []
-    if mode == 'soft':
-        exist_files = local_doc_qa.milvus_summary.check_file_exist_by_name(user_id, kb_id, file_names)
-        exist_file_names = [f[1] for f in exist_files]
-        for exist_file in exist_files:
-            file_id, file_name, file_size, status = exist_file
-            debug_logger.info(f"{url}, {status}, existed files, skip upload")
-            # await post_data(user_id, -1, file_id, status, msg='existed files, skip upload')
-    now = datetime.now()
-    timestamp = now.strftime("%Y%m%d%H%M")
-
-    data = []
-    for url, file_name in zip(urls, file_names):
-        if file_name in exist_file_names:
-            continue
-        local_file = LocalFile(user_id, kb_id, url, file_name)
-        file_id = local_file.file_id
-        file_size = len(local_file.file_content)
-        file_location = local_file.file_location
-        msg = local_doc_qa.milvus_summary.add_file(file_id, user_id, kb_id, file_name, file_size, file_location,
-                                                   chunk_size, timestamp, url)
-        debug_logger.info(f"{url}, {file_name}, {file_id}, {msg}")
-        data.append({"file_id": file_id, "file_name": file_name, "file_url": url, "status": "gray", "bytes": 0,
-                     "timestamp": timestamp})
-        # asyncio.create_task(local_doc_qa.insert_files_to_milvus(user_id, kb_id, [local_file]))
-    if exist_file_names:
-        msg = f'warning，当前的mode是soft，无法上传同名文件{exist_file_names}，如果想强制上传同名文件，请设置mode：strong'
-    else:
-        msg = "success，后台正在飞速上传文件，请耐心等待"
-    return sanic_json({"code": 200, "msg": msg, "data": data})
-
-
-@get_time_async
-@auth_required("write", check_kb_access=True)
-async def upload_files(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("upload_files %s", user_id)
-    debug_logger.info("user_info %s", user_info)
-    kb_id = safe_get(req, 'kb_id')
-    kb_id = correct_kb_id(kb_id)
-    debug_logger.info("kb_id %s", kb_id)
-    mode = safe_get(req, 'mode', default='soft')  # soft代表不上传同名文件，strong表示强制上传同名文件
-    debug_logger.info("mode: %s", mode)
-    chunk_size = safe_get(req, 'chunk_size', default=DEFAULT_PARENT_CHUNK_SIZE)
-    debug_logger.info("chunk_size: %s", chunk_size)
-    use_local_file = safe_get(req, 'use_local_file', 'false')
-    if use_local_file == 'true':
-        files = read_files_with_extensions()
-    else:
-        files = req.files.getlist('files')
-    debug_logger.info(f"{user_id} upload files number: {len(files)}")
-    not_exist_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, [kb_id])
-    if not_exist_kb_ids:
-        msg = "invalid kb_id: {}, please check...".format(not_exist_kb_ids)
-        return sanic_json({"code": 2001, "msg": msg, "data": [{}]})
-
-    exist_files = local_doc_qa.milvus_summary.get_files(user_id, kb_id)
-    if len(exist_files) + len(files) > 10000:
-        return sanic_json({"code": 2002,
-                           "msg": f"fail, exist files is {len(exist_files)}, upload files is {len(files)}, total files is {len(exist_files) + len(files)}, max length is 10000."})
-
-    data = []
-    local_files = []
-    file_names = []
-    for file in files:
-        if isinstance(file, str):
-            file_name = os.path.basename(file)
-        else:
-            debug_logger.info('ori name: %s', file.name)
-            file_name = urllib.parse.unquote(file.name, encoding='UTF-8')
-            debug_logger.info('decode name: %s', file_name)
-        # # 使用正则表达式替换以%开头的字符串
-        # file_name = re.sub(r'%\w+', '', file_name)
-        # 删除掉全角字符
-        file_name = re.sub(r'[\uFF01-\uFF5E\u3000-\u303F]', '', file_name)
-        debug_logger.info('cleaned name: %s', file_name)
-        # max_length = 255 - len(construct_qanything_local_file_nos_key_prefix(file_id)) == 188
-        file_name = truncate_filename(file_name, max_length=110)
-        file_names.append(file_name)
-
-    exist_file_names = []
-    if mode == 'soft':
-        exist_files = local_doc_qa.milvus_summary.check_file_exist_by_name(user_id, kb_id, file_names)
-        exist_file_names = [f[1] for f in exist_files]
-        for exist_file in exist_files:
-            file_id, file_name, file_size, status = exist_file
-            debug_logger.info(f"{file_name}, {status}, existed files, skip upload")
-            # await post_data(user_id, -1, file_id, status, msg='existed files, skip upload')
-
-    now = datetime.now()
-    timestamp = now.strftime("%Y%m%d%H%M")
-
-    failed_files = []
-    for file, file_name in zip(files, file_names):
-        if file_name in exist_file_names:
-            continue
-        local_file = LocalFile(user_id, kb_id, file, file_name)
-        chars = fast_estimate_file_char_count(local_file.file_location)
-        debug_logger.info(f"{file_name} char_size: {chars}")
-        if chars and chars > MAX_CHARS:
-            debug_logger.warning(f"fail, file {file_name} chars is {chars}, max length is {MAX_CHARS}.")
-            # return sanic_json({"code": 2003, "msg": f"fail, file {file_name} chars is too much, max length is {MAX_CHARS}."})
-            failed_files.append(file_name)
-            continue
-        file_id = local_file.file_id
-        file_size = len(local_file.file_content)
-        file_location = local_file.file_location
-        local_files.append(local_file)
-        msg = local_doc_qa.milvus_summary.add_file(file_id, user_id, kb_id, file_name, file_size, file_location,
-                                                   chunk_size, timestamp)
-        debug_logger.info(f"{file_name}, {file_id}, {msg}")
-        data.append(
-            {"file_id": file_id, "file_name": file_name, "status": "gray", "bytes": len(local_file.file_content),
-             "timestamp": timestamp, "estimated_chars": chars})
-
-    # asyncio.create_task(local_doc_qa.insert_files_to_milvus(user_id, kb_id, local_files))
-    if exist_file_names:
-        msg = f'warning，当前的mode是soft，无法上传同名文件{exist_file_names}，如果想强制上传同名文件，请设置mode：strong'
-    elif failed_files:
-        msg = f"warning, {failed_files} chars is too much, max characters length is {MAX_CHARS}, skip upload."
-    else:
-        msg = "success，后台正在飞速上传文件，请耐心等待"
-    return sanic_json({"code": 200, "msg": msg, "data": data})
-
-
-@get_time_async
-async def upload_faqs(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("upload_faqs %s", user_id)
-    debug_logger.info("user_info %s", user_info)
-    kb_id = safe_get(req, 'kb_id')
-    kb_id = correct_kb_id(kb_id)
-    debug_logger.info("kb_id %s", kb_id)
-    faqs = safe_get(req, 'faqs')
-    chunk_size = safe_get(req, 'chunk_size', default=DEFAULT_PARENT_CHUNK_SIZE)
-    debug_logger.info("chunk_size: %s", chunk_size)
-
-    # 增加上传文件的功能和上传文件的检查解析
-    file_status = {}
-    if faqs is None:
-        files = req.files.getlist('files')
-        faqs = []
-        for file in files:
-            debug_logger.info('ori name: %s', file.name)
-            file_name = urllib.parse.unquote(file.name, encoding='UTF-8')
-            debug_logger.info('decode name: %s', file_name)
-            # 删除掉全角字符
-            file_name = re.sub(r'[\uFF01-\uFF5E\u3000-\u303F]', '', file_name)
-            file_name = file_name.replace("/", "_")
-            debug_logger.info('cleaned name: %s', file_name)
-            file_name = truncate_filename(file_name)
-            file_faqs = check_and_transform_excel(file.body)
-            if isinstance(file_faqs, str):
-                file_status[file_name] = file_faqs
-            else:
-                faqs.extend(file_faqs)
-                file_status[file_name] = "success"
-
-    if len(faqs) > 1000:
-        return sanic_json({"code": 2002, "msg": f"fail, faqs too many, max length is 1000."})
-
-    not_exist_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, [kb_id])
-    if not_exist_kb_ids:
-        msg = "invalid kb_id: {}, please check...".format(not_exist_kb_ids)
-        return sanic_json({"code": 2001, "msg": msg})
-
-    data = []
-    local_files = []
-    now = datetime.now()
-    timestamp = now.strftime("%Y%m%d%H%M")
-    debug_logger.info(f"start insert {len(faqs)} faqs to mysql, user_id: {user_id}, kb_id: {kb_id}")
-    for faq in tqdm(faqs):
-        ques = faq['question']
-        if len(ques) > 512 or len(faq['answer']) > 2048:
-            return sanic_json(
-                {"code": 2003, "msg": f"fail, faq too long, max length of question is 512, answer is 2048."})
-        file_name = f"FAQ_{ques}.faq"
-        file_name = file_name.replace("/", "_").replace(":", "_")  # 文件名中的/和：会导致写入时出错
-        file_name = simplify_filename(file_name)
-        file_size = len(ques) + len(faq['answer'])
-        # faq_id = local_doc_qa.milvus_summary.get_faq_by_question(ques, kb_id)
-        # if faq_id:
-        #     debug_logger.info(f"faq question {ques} already exist, skip")
-        #     data.append({
-        #         "file_id": faq_id,
-        #         "file_name": file_name,
-        #         "status": "green",
-        #         "length": file_size,
-        #         "timestamp": local_doc_qa.milvus_summary.get_file_timestamp(faq_id)
-        #     })
-        #     continue
-        local_file = LocalFile(user_id, kb_id, faq, file_name)
-        file_id = local_file.file_id
-        file_location = local_file.file_location
-        local_files.append(local_file)
-        local_doc_qa.milvus_summary.add_faq(file_id, user_id, kb_id, faq['question'], faq['answer'], faq.get('nos_keys', ''))
-        local_doc_qa.milvus_summary.add_file(file_id, user_id, kb_id, file_name, file_size, file_location,
-                                             chunk_size, timestamp)
-        # debug_logger.info(f"{file_name}, {file_id}, {msg}, {faq}")
-        data.append(
-            {"file_id": file_id, "file_name": file_name, "status": "gray", "length": file_size,
-             "timestamp": timestamp})
-    debug_logger.info(f"end insert {len(faqs)} faqs to mysql, user_id: {user_id}, kb_id: {kb_id}")
-
-    msg = "success，后台正在飞速上传文件，请耐心等待"
-    return sanic_json({"code": 200, "msg": msg, "data": data})
-
-
-@get_time_async
-@auth_required("read", check_kb_access=False)
-async def list_kbs(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("list_kbs %s", user_id)
-    kb_infos = local_doc_qa.milvus_summary.get_knowledge_bases(user_id)
-    data = []
-    for kb in kb_infos:
-        data.append({"kb_id": kb[0], "kb_name": kb[1]})
-    debug_logger.info("all kb infos: {}".format(data))
-    return sanic_json({"code": 200, "data": data})
-
-
-@get_time_async
-@auth_required("read", check_kb_access=True)
-async def list_docs(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("list_docs %s", user_id)
-    kb_id = safe_get(req, 'kb_id')
-    kb_id = correct_kb_id(kb_id)
-    debug_logger.info("kb_id: {}".format(kb_id))
-    file_id = safe_get(req, 'file_id')
-    page_id = safe_get(req, 'page_id', 1)  # 默认为第一页
-    page_limit = safe_get(req, 'page_limit', 10)  # 默认每页显示10条记录
-    data = []
-    if file_id is None:
-        file_infos = local_doc_qa.milvus_summary.get_files(user_id, kb_id)
-    else:
-        file_infos = local_doc_qa.milvus_summary.get_files(user_id, kb_id, file_id)
-    status_count = {}
-    # msg_map = {'gray': "已上传到服务器，进入上传等待队列",
-    #            'red': "上传出错，请删除后重试或联系工作人员",
-    #            'yellow': "已进入上传队列，请耐心等待", 'green': "上传成功"}
-    for file_info in file_infos:
-        status = file_info[2]
-        if status not in status_count:
-            status_count[status] = 1
-        else:
-            status_count[status] += 1
-        data.append({"file_id": file_info[0], "file_name": file_info[1], "status": file_info[2], "bytes": file_info[3],
-                     "content_length": file_info[4], "timestamp": file_info[5], "file_location": file_info[6],
-                     "file_url": file_info[7], "chunks_number": file_info[8], "msg": file_info[9]})
-        if file_info[1].endswith('.faq'):
-            faq_info = local_doc_qa.milvus_summary.get_faq(file_info[0])
-            user_id, kb_id, question, answer, nos_keys = faq_info
-            data[-1]['question'] = question
-            data[-1]['answer'] = answer
-
-    # data根据timestamp排序，时间越新的越靠前
-    data = sorted(data, key=lambda x: int(x['timestamp']), reverse=True)
-
-    # 计算总记录数
-    total_count = len(data)
-    # 计算总页数
-    total_pages = (total_count + page_limit - 1) // page_limit
-    if page_id > total_pages and total_count != 0:
-        return sanic_json({"code": 2002, "msg": f'输入非法！page_id超过最大值，page_id: {page_id}，最大值：{total_pages}，请检查！'})
-    # 计算当前页的起始和结束索引
-    start_index = (page_id - 1) * page_limit
-    end_index = start_index + page_limit
-    # 截取当前页的数据
-    current_page_data = data[start_index:end_index]
-
-    # return sanic_json({"code": 200, "msg": "success", "data": {'total': status_count, 'details': data}})
-    return sanic_json({
-        "code": 200,
-        "msg": "success",
-        "data": {
-            'total_page': total_pages,  # 总页数
-            "total": total_count,  # 总文件数
-            "status_count": status_count,  # 各状态的文件数
-            "details": current_page_data,  # 当前页码下的文件目录
-            "page_id": page_id,  # 当前页码,
-            "page_limit": page_limit  # 每页显示的文件数
-        }
-    })
-
-
-@get_time_async
-@auth_required("admin", check_kb_access=True)
-async def delete_knowledge_base(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    # TODO: 确认是否支持批量删除知识库
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("delete_knowledge_base %s", user_id)
-    kb_ids = safe_get(req, 'kb_ids')
-    kb_ids = [correct_kb_id(kb_id) for kb_id in kb_ids]
-    not_exist_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, kb_ids)
-    if not_exist_kb_ids:
-        return sanic_json({"code": 2003, "msg": "fail, knowledge Base {} not found".format(not_exist_kb_ids)})
-
-    # milvus = local_doc_qa.match_milvus_kb(user_id, kb_ids)
-    for kb_id in kb_ids:
-        expr = f"kb_id == \"{kb_id}\""
-        asyncio.create_task(run_in_background(local_doc_qa.milvus_kb.delete_expr, expr))
-        # local_doc_qa.milvus_kb.delete_expr(expr)
-        # milvus.delete_partition(kb_id)
-    for kb_id in kb_ids:
-        file_infos = local_doc_qa.milvus_summary.get_files(user_id, kb_id)
-        file_ids = [file_info[0] for file_info in file_infos]
-        file_chunks = [file_info[8] for file_info in file_infos]
-        asyncio.create_task(run_in_background(local_doc_qa.es_client.delete_files, file_ids, file_chunks))
-        local_doc_qa.milvus_summary.delete_documents(file_ids)
-        local_doc_qa.milvus_summary.delete_faqs(file_ids)
-
-        # delete kb_id file dir
-        try:
-            upload_path = os.path.join(UPLOAD_ROOT_PATH, user_id)
-            file_dir = os.path.join(upload_path, kb_id)
-            debug_logger.info("delete_knowledge_base file dir : %s", file_dir)
-            shutil.rmtree(file_dir)
-        except Exception as e:
-            debug_logger.error("An error occurred while constructing file paths: %s", str(e))
-
-
-        debug_logger.info(f"""delete knowledge base {kb_id} success""")
-    local_doc_qa.milvus_summary.delete_knowledge_base(user_id, kb_ids)
-    return sanic_json({"code": 200, "msg": "Knowledge Base {} delete success".format(kb_ids)})
-
-
-@get_time_async
-@auth_required("write", check_kb_access=True)
-async def rename_knowledge_base(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("rename_knowledge_base %s", user_id)
-    kb_id = safe_get(req, 'kb_id')
-    kb_id = correct_kb_id(kb_id)
-    new_kb_name = safe_get(req, 'new_kb_name')
-    not_exist_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, [kb_id])
-    if not_exist_kb_ids:
-        return sanic_json({"code": 2003, "msg": "fail, knowledge Base {} not found".format(not_exist_kb_ids[0])})
-    local_doc_qa.milvus_summary.rename_knowledge_base(user_id, kb_id, new_kb_name)
-    return sanic_json({"code": 200, "msg": "Knowledge Base {} rename success".format(kb_id)})
-
-
-@get_time_async
-@auth_required("write", check_kb_access=True)
-async def delete_docs(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("delete_docs %s", user_id)
-    kb_id = safe_get(req, 'kb_id')
-    kb_id = correct_kb_id(kb_id)
-    file_ids = safe_get(req, "file_ids")
-    not_exist_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, [kb_id])
-    if not_exist_kb_ids:
-        return sanic_json({"code": 2003, "msg": "fail, knowledge Base {} not found".format(not_exist_kb_ids[0])})
-    valid_file_infos = local_doc_qa.milvus_summary.check_file_exist(user_id, kb_id, file_ids)
-    if len(valid_file_infos) == 0:
-        return sanic_json({"code": 2004, "msg": "fail, files {} not found".format(file_ids)})
-    valid_file_ids = [file_info[0] for file_info in valid_file_infos]
-    debug_logger.info("delete_docs valid_file_ids %s", valid_file_ids)
-    # milvus_kb = local_doc_qa.match_milvus_kb(user_id, [kb_id])
-    # milvus_kb.delete_files(file_ids)
-    expr = f"""kb_id == "{kb_id}" and file_id in {valid_file_ids}"""  # 删除数据库中的记录
-    asyncio.create_task(run_in_background(local_doc_qa.milvus_kb.delete_expr, expr))
-    # local_doc_qa.milvus_kb.delete_expr(expr)
-    file_chunks = local_doc_qa.milvus_summary.get_chunk_size(valid_file_ids)
-    asyncio.create_task(run_in_background(local_doc_qa.es_client.delete_files, valid_file_ids, file_chunks))
-
-    local_doc_qa.milvus_summary.delete_files(kb_id, valid_file_ids)
-    local_doc_qa.milvus_summary.delete_documents(valid_file_ids)
-    local_doc_qa.milvus_summary.delete_faqs(valid_file_ids)
-    # list file_ids
-    for file_id in file_ids:
-        try:
-            upload_path = os.path.join(UPLOAD_ROOT_PATH, user_id)
-            file_dir = os.path.join(upload_path, kb_id, file_id)
-            debug_logger.info("delete_docs file_dir %s", file_dir)
-            # delete file dir
-            shutil.rmtree(file_dir)
-            # delele images dir
-            images_dir = os.path.join(IMAGES_ROOT_PATH, file_id)
-            debug_logger.info("delete_docs images_dir %s", images_dir)
-            shutil.rmtree(images_dir)
-        except Exception as e:
-            debug_logger.error("An error occurred while constructing file paths: %s", str(e))
-
-    return sanic_json({"code": 200, "msg": "documents {} delete success".format(valid_file_ids)})
-
-
-@get_time_async
-async def get_total_status(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info('get_total_status %s', user_id)
-    by_date = safe_get(req, 'by_date', False)
-    if not user_id:
-        users = local_doc_qa.milvus_summary.get_users()
-        users = [user[0] for user in users]
-    else:
-        users = [user_id]
-    res = {}
-    for user in users:
-        res[user] = {}
-        if by_date:
-            res[user] = local_doc_qa.milvus_summary.get_total_status_by_date(user)
-            continue
-        kbs = local_doc_qa.milvus_summary.get_knowledge_bases(user)
-        for kb_id, kb_name in kbs:
-            gray_file_infos = local_doc_qa.milvus_summary.get_file_by_status([kb_id], 'gray')
-            red_file_infos = local_doc_qa.milvus_summary.get_file_by_status([kb_id], 'red')
-            yellow_file_infos = local_doc_qa.milvus_summary.get_file_by_status([kb_id], 'yellow')
-            green_file_infos = local_doc_qa.milvus_summary.get_file_by_status([kb_id], 'green')
-            res[user][kb_name + kb_id] = {'green': len(green_file_infos), 'yellow': len(yellow_file_infos),
-                                          'red': len(red_file_infos),
-                                          'gray': len(gray_file_infos)}
-
-    return sanic_json({"code": 200, "status": res})
-
-
-@get_time_async
-async def clean_files_by_status(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info('clean_files_by_status %s', user_id)
-    status = safe_get(req, 'status', default='gray')
-    if status not in ['gray', 'red', 'yellow']:
-        return sanic_json({"code": 2003, "msg": "fail, status {} must be in ['gray', 'red', 'yellow']".format(status)})
-    kb_ids = safe_get(req, 'kb_ids')
-    kb_ids = [correct_kb_id(kb_id) for kb_id in kb_ids]
-    if not kb_ids:
-        kbs = local_doc_qa.milvus_summary.get_knowledge_bases(user_id)
-        kb_ids = [kb[0] for kb in kbs]
-    else:
-        not_exist_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, kb_ids)
-        if not_exist_kb_ids:
-            return sanic_json({"code": 2003, "msg": "fail, knowledge Base {} not found".format(not_exist_kb_ids)})
-
-    gray_file_infos = local_doc_qa.milvus_summary.get_file_by_status(kb_ids, status)
-    gray_file_ids = [f[0] for f in gray_file_infos]
-    gray_file_names = [f[1] for f in gray_file_infos]
-    debug_logger.info(f'{status} files number: {len(gray_file_names)}')
-    # 删除milvus中的file
-    if gray_file_ids:
-        # expr = f"file_id in \"{gray_file_ids}\""
-        # asyncio.create_task(run_in_background(local_doc_qa.milvus_kb.delete_expr, expr))
-        for kb_id in kb_ids:
-            local_doc_qa.milvus_summary.delete_files(kb_id, gray_file_ids)
-    return sanic_json({"code": 200, "msg": f"delete {status} files success", "data": gray_file_names})
-
-
-@get_time_async
-async def local_doc_chat(req: request):
-    preprocess_start = time.perf_counter()
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # local_cluster = get_milvus_cluster_by_user_info(user_info)
-    # user_id = user_id + '__' + user_info
-    # local_doc_qa.milvus_summary.update_user_cluster(user_id, [get_milvus_cluster_by_user_info(user_info)])
-    debug_logger.info('local_doc_chat %s', user_id)
-    debug_logger.info('user_info %s', user_info)
-    bot_id = safe_get(req, 'bot_id')
-    if bot_id:
-        if not local_doc_qa.milvus_summary.check_bot_is_exist(bot_id):
-            return sanic_json({"code": 2003, "msg": "fail, Bot {} not found".format(bot_id)})
-        bot_info = local_doc_qa.milvus_summary.get_bot(None, bot_id)[0]
-        bot_id, bot_name, desc, image, prompt, welcome, kb_ids_str, upload_time, user_id, llm_setting = bot_info
-        kb_ids = kb_ids_str.split(',')
-        if not kb_ids:
-            return sanic_json({"code": 2003, "msg": "fail, Bot {} unbound knowledge base.".format(bot_id)})
-        custom_prompt = prompt
-        if not llm_setting:
-            return sanic_json({"code": 2003, "msg": "fail, Bot {} llm_setting is empty.".format(bot_id)})
-        llm_setting = json.loads(llm_setting)
-        rerank = llm_setting.get('rerank', True)
-        only_need_search_results = llm_setting.get('only_need_search_results', False)
-        need_web_search = llm_setting.get('networking', False)
-        api_base = llm_setting.get('api_base', '')
-        api_key = llm_setting.get('api_key', 'ollama')
-        api_context_length = llm_setting.get('api_context_length', 4096)
-        top_p = llm_setting.get('top_p', 0.99)
-        temperature = llm_setting.get('temperature', 0.5)
-        top_k = llm_setting.get('top_k', VECTOR_SEARCH_TOP_K)
-        model = llm_setting.get('model', 'gpt-4o-mini')
-        max_token = llm_setting.get('max_token')
-        hybrid_search = llm_setting.get('hybrid_search', False)
-        chunk_size = llm_setting.get('chunk_size', DEFAULT_PARENT_CHUNK_SIZE)
-    else:
-        kb_ids = safe_get(req, 'kb_ids')
-        custom_prompt = safe_get(req, 'custom_prompt', None)
-        rerank = safe_get(req, 'rerank', default=True)
-        only_need_search_results = safe_get(req, 'only_need_search_results', False)
-        need_web_search = safe_get(req, 'networking', False)
-        api_base = safe_get(req, 'api_base', '')
-        # 如果api_base中包含0.0.0.0或127.0.0.1或localhost，替换为GATEWAY_IP
-        api_base = api_base.replace('0.0.0.0', GATEWAY_IP).replace('127.0.0.1', GATEWAY_IP).replace('localhost',
-                                                                                                    GATEWAY_IP)
-        api_key = safe_get(req, 'api_key', 'ollama')
-        api_context_length = safe_get(req, 'api_context_length', 4096)
-        top_p = safe_get(req, 'top_p', 0.99)
-        temperature = safe_get(req, 'temperature', 0.5)
-        top_k = safe_get(req, 'top_k', VECTOR_SEARCH_TOP_K)
-
-        model = safe_get(req, 'model', 'gpt-4o-mini')
-        max_token = safe_get(req, 'max_token')
-
-        hybrid_search = safe_get(req, 'hybrid_search', False)
-        chunk_size = safe_get(req, 'chunk_size', DEFAULT_PARENT_CHUNK_SIZE)
-
-    debug_logger.info('rerank %s', rerank)
-
-    if len(kb_ids) > 20:
-        return sanic_json({"code": 2005, "msg": "fail, kb_ids length should less than or equal to 20"})
-    kb_ids = [correct_kb_id(kb_id) for kb_id in kb_ids]
-    question = safe_get(req, 'question')
-    streaming = safe_get(req, 'streaming', False)
-    history = safe_get(req, 'history', [])
-
-    if top_k > 100:
-        return sanic_json({"code": 2003, "msg": "fail, top_k should less than or equal to 100"})
-
-    missing_params = []
-    if not api_base:
-        missing_params.append('api_base')
-    if not api_key:
-        missing_params.append('api_key')
-    if not api_context_length:
-        missing_params.append('api_context_length')
-    if not top_p:
-        missing_params.append('top_p')
-    if not top_k:
-        missing_params.append('top_k')
-    if top_p == 1.0:
-        top_p = 0.99
-    if not temperature:
-        missing_params.append('temperature')
-
-    if missing_params:
-        missing_params_str = " and ".join(missing_params) if len(missing_params) > 1 else missing_params[0]
-        return sanic_json({"code": 2003, "msg": f"fail, {missing_params_str} is required"})
-
-    if only_need_search_results and streaming:
-        return sanic_json(
-            {"code": 2006, "msg": "fail, only_need_search_results and streaming can't be True at the same time"})
-    request_source = safe_get(req, 'source', 'unknown')
-
-    debug_logger.info("history: %s ", history)
-    debug_logger.info("question: %s", question)
-    debug_logger.info("kb_ids: %s", kb_ids)
-    debug_logger.info("user_id: %s", user_id)
-    debug_logger.info("custom_prompt: %s", custom_prompt)
-    debug_logger.info("model: %s", model)
-    debug_logger.info("max_token: %s", max_token)
-    debug_logger.info("request_source: %s", request_source)
-    debug_logger.info("only_need_search_results: %s", only_need_search_results)
-    debug_logger.info("bot_id: %s", bot_id)
-    debug_logger.info("need_web_search: %s", need_web_search)
-    debug_logger.info("api_base: %s", api_base)
-    debug_logger.info("api_key: %s", api_key)
-    debug_logger.info("api_context_length: %s", api_context_length)
-    debug_logger.info("top_p: %s", top_p)
-    debug_logger.info("top_k: %s", top_k)
-    debug_logger.info("temperature: %s", temperature)
-    debug_logger.info("hybrid_search: %s", hybrid_search)
-    debug_logger.info("chunk_size: %s", chunk_size)
-
-    time_record = {}
-    if kb_ids:
-        not_exist_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, kb_ids)
-        if not_exist_kb_ids:
-            return sanic_json({"code": 2003, "msg": "fail, knowledge Base {} not found".format(not_exist_kb_ids)})
-        faq_kb_ids = [kb + '_FAQ' for kb in kb_ids]
-        not_exist_faq_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, faq_kb_ids)
-        exist_faq_kb_ids = [kb for kb in faq_kb_ids if kb not in not_exist_faq_kb_ids]
-        debug_logger.info("exist_faq_kb_ids: %s", exist_faq_kb_ids)
-        kb_ids += exist_faq_kb_ids
-
-    file_infos = []
-    for kb_id in kb_ids:
-        file_infos.extend(local_doc_qa.milvus_summary.get_files(user_id, kb_id))
-    valid_files = [fi for fi in file_infos if fi[2] == 'green']
-    if len(valid_files) == 0:
-        debug_logger.info("valid_files is empty, use only chat mode.")
-        kb_ids = []
-    preprocess_end = time.perf_counter()
-    time_record['preprocess'] = round(preprocess_end - preprocess_start, 2)
-    # 获取格式为'2021-08-01 00:00:00'的时间戳
-    qa_timestamp = time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(time.time()))
-    for kb_id in kb_ids:
-        local_doc_qa.milvus_summary.update_knowledge_base_latest_qa_time(kb_id, qa_timestamp)
-    debug_logger.info("streaming: %s", streaming)
-    if streaming:
-        debug_logger.info("start generate answer")
-
-        async def generate_answer(response):
-            debug_logger.info("start generate...")
-            async for resp, next_history in local_doc_qa.get_knowledge_based_answer(model=model,
-                                                                                    max_token=max_token,
-                                                                                    kb_ids=kb_ids,
-                                                                                    query=question,
-                                                                                    retriever=local_doc_qa.retriever,
-                                                                                    chat_history=history,
-                                                                                    streaming=True,
-                                                                                    rerank=rerank,
-                                                                                    custom_prompt=custom_prompt,
-                                                                                    time_record=time_record,
-                                                                                    need_web_search=need_web_search,
-                                                                                    hybrid_search=hybrid_search,
-                                                                                    web_chunk_size=chunk_size,
-                                                                                    temperature=temperature,
-                                                                                    api_base=api_base,
-                                                                                    api_key=api_key,
-                                                                                    api_context_length=api_context_length,
-                                                                                    top_p=top_p,
-                                                                                    top_k=top_k
-                                                                                    ):
-                chunk_data = resp["result"]
-                if not chunk_data:
-                    continue
-                chunk_str = chunk_data[6:]
-                if chunk_str.startswith("[DONE]"):
-                    retrieval_documents = format_source_documents(resp["retrieval_documents"])
-                    source_documents = format_source_documents(resp["source_documents"])
-                    result = next_history[-1][1]
-                    # result = resp['result']
-                    time_record['chat_completed'] = round(time.perf_counter() - preprocess_start, 2)
-                    if time_record.get('llm_completed', 0) > 0:
-                        time_record['tokens_per_second'] = round(
-                            len(result) / time_record['llm_completed'], 2)
-                    formatted_time_record = format_time_record(time_record)
-                    chat_data = {'user_id': user_id, 'kb_ids': kb_ids, 'query': question, "model": model,
-                                 "product_source": request_source, 'time_record': formatted_time_record,
-                                 'history': history,
-                                 'condense_question': resp['condense_question'], 'prompt': resp['prompt'],
-                                 'result': result, 'retrieval_documents': retrieval_documents,
-                                 'source_documents': source_documents, 'bot_id': bot_id}
-                    local_doc_qa.milvus_summary.add_qalog(**chat_data)
-                    qa_logger.info("chat_data: %s", chat_data)
-                    debug_logger.info("response: %s", chat_data['result'])
-                    stream_res = {
-                        "code": 200,
-                        "msg": "success stream chat",
-                        "question": question,
-                        "response": result,
-                        "model": model,
-                        "history": next_history,
-                        "condense_question": resp['condense_question'],
-                        "source_documents": source_documents,
-                        "retrieval_documents": retrieval_documents,
-                        "time_record": formatted_time_record,
-                        "show_images": resp.get('show_images', [])
-                    }
-                else:
-                    time_record['rollback_length'] = resp.get('rollback_length', 0)
-                    if 'first_return' not in time_record:
-                        time_record['first_return'] = round(time.perf_counter() - preprocess_start, 2)
-                    chunk_js = json.loads(chunk_str)
-                    delta_answer = chunk_js["answer"]
-                    stream_res = {
-                        "code": 200,
-                        "msg": "success",
-                        "question": "",
-                        "response": delta_answer,
-                        "history": [],
-                        "source_documents": [],
-                        "retrieval_documents": [],
-                        "time_record": format_time_record(time_record),
-                    }
-                await response.write(f"data: {json.dumps(stream_res, ensure_ascii=False)}\n\n")
-                if chunk_str.startswith("[DONE]"):
-                    await response.eof()
-                await asyncio.sleep(0.001)
-
-        response_stream = ResponseStream(generate_answer, content_type='text/event-stream')
-        return response_stream
-
-    else:
-        async for resp, history in local_doc_qa.get_knowledge_based_answer(model=model,
-                                                                           max_token=max_token,
-                                                                           kb_ids=kb_ids,
-                                                                           query=question,
-                                                                           retriever=local_doc_qa.retriever,
-                                                                           chat_history=history, streaming=False,
-                                                                           rerank=rerank,
-                                                                           custom_prompt=custom_prompt,
-                                                                           time_record=time_record,
-                                                                           only_need_search_results=only_need_search_results,
-                                                                           need_web_search=need_web_search,
-                                                                           hybrid_search=hybrid_search,
-                                                                           web_chunk_size=chunk_size,
-                                                                           temperature=temperature,
-                                                                           api_base=api_base,
-                                                                           api_key=api_key,
-                                                                           api_context_length=api_context_length,
-                                                                           top_p=top_p,
-                                                                           top_k=top_k
-                                                                           ):
-            pass
-        if only_need_search_results:
-            return sanic_json(
-                {"code": 200, "question": question, "source_documents": format_source_documents(resp)})
-        retrieval_documents = format_source_documents(resp["retrieval_documents"])
-        source_documents = format_source_documents(resp["source_documents"])
-        formatted_time_record = format_time_record(time_record)
-        chat_data = {'user_id': user_id, 'kb_ids': kb_ids, 'query': question, 'time_record': formatted_time_record,
-                     'history': history, "condense_question": resp['condense_question'], "model": model,
-                     "product_source": request_source,
-                     'retrieval_documents': retrieval_documents, 'prompt': resp['prompt'], 'result': resp['result'],
-                     'source_documents': source_documents, 'bot_id': bot_id}
-        local_doc_qa.milvus_summary.add_qalog(**chat_data)
-        qa_logger.info("chat_data: %s", chat_data)
-        debug_logger.info("response: %s", chat_data['result'])
-        return sanic_json({"code": 200, "msg": "success no stream chat", "question": question,
-                           "response": resp["result"], "model": model,
-                           "history": history, "condense_question": resp['condense_question'],
-                           "source_documents": source_documents, "retrieval_documents": retrieval_documents,
-                           "time_record": formatted_time_record})
-
 
 @get_time_async
 async def document(req: request):
@@ -1082,644 +131,3 @@ https://qanything.youdao.com
 
 """
     return sanic_text(description)
-
-
-@get_time_async
-async def get_doc_completed(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("get_doc_chunks %s", user_id)
-    kb_id = safe_get(req, 'kb_id')
-    kb_id = correct_kb_id(kb_id)
-    debug_logger.info("kb_id: {}".format(kb_id))
-    file_id = safe_get(req, 'file_id')
-    if not file_id:
-        return sanic_json({"code": 2005, "msg": "fail, file_id is None"})
-    debug_logger.info("file_id: {}".format(file_id))
-    page_id = safe_get(req, 'page_id', 1)  # 默认为第一页
-    page_limit = safe_get(req, 'page_limit', 10)  # 默认每页显示10条记录
-
-    sorted_json_datas = local_doc_qa.milvus_summary.get_document_by_file_id(file_id)
-    # completed_doc = local_doc_qa.get_completed_document(file_id)
-    # for json_data in sorted_json_datas:
-    #     completed_text += json_data['kwargs']['page_content'] + '\n'
-    #     if len(completed_text) > 10000:
-    #         return sanic_json({"code": 200, "msg": "failed, completed_text too long, the max length is 10000"})
-    chunks = [json_data['kwargs'] for json_data in sorted_json_datas]
-
-    # 计算总记录数
-    total_count = len(chunks)
-    # 计算总页数
-    total_pages = (total_count + page_limit - 1) // page_limit
-    if page_id > total_pages and total_count != 0:
-        return sanic_json({"code": 2002, "msg": f'输入非法！page_id超过最大值，page_id: {page_id}，最大值：{total_pages}，请检查！'})
-    # 计算当前页的起始和结束索引
-    start_index = (page_id - 1) * page_limit
-    end_index = start_index + page_limit
-    # 截取当前页的数据
-    current_page_chunks = chunks[start_index:end_index]
-    for chunk in current_page_chunks:
-        chunk['page_content'] = replace_image_references(chunk['page_content'], file_id)
-
-    # return sanic_json({"code": 200, "msg": "success", "completed_text": completed_doc.page_content,
-    #                    "chunks": current_page_chunks, "page_id": page, "total_count": total_count})
-    file_location = local_doc_qa.milvus_summary.get_file_location(file_id)
-    # 获取file_location的上一级目录
-    file_path = os.path.dirname(file_location)
-    return sanic_json({"code": 200, "msg": "success", "chunks": current_page_chunks, "file_path": file_path,
-                       "page_id": page_id, "page_limit": page_limit, "total_count": total_count})
-
-
-@get_time_async
-async def get_qa_info(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    any_kb_id = safe_get(req, 'any_kb_id')
-    user_id = safe_get(req, 'user_id')
-    if user_id is None and not any_kb_id:
-        return sanic_json({"code": 2005, "msg": "fail, user_id and any_kb_id is None"})
-    if any_kb_id:
-        any_kb_id = correct_kb_id(any_kb_id)
-        debug_logger.info("get_qa_info %s", any_kb_id)
-    if user_id:
-        user_info = safe_get(req, 'user_info', "1234")
-        passed, msg = check_user_id_and_user_info(user_id, user_info)
-        if not passed:
-            return sanic_json({"code": 2001, "msg": msg})
-        # user_id = user_id + '__' + user_info
-        debug_logger.info("get_qa_info %s", user_id)
-    query = safe_get(req, 'query')
-    bot_id = safe_get(req, 'bot_id')
-    qa_ids = safe_get(req, "qa_ids")
-    time_start = safe_get(req, 'time_start')
-    time_end = safe_get(req, 'time_end')
-    time_range = get_time_range(time_start, time_end)
-    if not time_range:
-        return {"code": 2002, "msg": f'输入非法！time_start格式错误，time_start: {time_start}，示例：2024-10-05，请检查！'}
-    only_need_count = safe_get(req, 'only_need_count', False)
-    debug_logger.info(f"only_need_count: {only_need_count}")
-    if only_need_count:
-        need_info = ["timestamp"]
-        qa_infos = local_doc_qa.milvus_summary.get_qalog_by_filter(need_info=need_info, user_id=user_id, time_range=time_range)
-        # timestamp = now.strftime("%Y%m%d%H%M")
-        # 按照timestamp，按照天数进行统计，比如20240628，20240629，20240630，计算每天的问答数量
-        qa_infos = sorted(qa_infos, key=lambda x: x['timestamp'])
-        qa_infos = [qa_info['timestamp'] for qa_info in qa_infos]
-        qa_infos = [qa_info[:10] for qa_info in qa_infos]
-        qa_infos_by_day = dict(Counter(qa_infos))
-        return sanic_json({"code": 200, "msg": "success", "qa_infos_by_day": qa_infos_by_day})
-
-    page_id = safe_get(req, 'page_id', 1)
-    page_limit = safe_get(req, 'page_limit', 10)
-    default_need_info = ["qa_id", "user_id", "bot_id", "kb_ids", "query", "model", "product_source", "time_record",
-                         "history", "condense_question", "prompt", "result", "retrieval_documents", "source_documents",
-                         "timestamp"]
-    need_info = safe_get(req, 'need_info', default_need_info)
-    save_to_excel = safe_get(req, 'save_to_excel', False)
-    qa_infos = local_doc_qa.milvus_summary.get_qalog_by_filter(need_info=need_info, user_id=user_id, query=query,
-                                                               bot_id=bot_id, time_range=time_range,
-                                                               any_kb_id=any_kb_id, qa_ids=qa_ids)
-    if save_to_excel:
-        timestamp = datetime.now().strftime("%Y%m%d%H%M")
-        file_name = f"QAnything_QA_{timestamp}.xlsx"
-        file_path = export_qalogs_to_excel(qa_infos, need_info, file_name)
-        return await response.file(file_path, filename=file_name,
-                                   mime_type='application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-                                   headers={'Content-Disposition': f'attachment; filename="{file_name}"'})
-
-    # 计算总记录数
-    total_count = len(qa_infos)
-    # 计算总页数
-    total_pages = (total_count + page_limit - 1) // page_limit
-    if page_id > total_pages and total_count != 0:
-        return sanic_json(
-            {"code": 2002, "msg": f'输入非法！page_id超过最大值，page_id: {page_id}，最大值：{total_pages}，请检查！'})
-    # 计算当前页的起始和结束索引
-    start_index = (page_id - 1) * page_limit
-    end_index = start_index + page_limit
-    # 截取当前页的数据
-    current_qa_infos = qa_infos[start_index:end_index]
-    msg = f"检测到的Log总数为{total_count}, 本次返回page_id为{page_id}的数据，每页显示{page_limit}条"
-
-    # if len(qa_infos) > 100:
-    #     pages = math.ceil(len(qa_infos) // 100)
-    #     if page_id is None:
-    #         msg = f"检索到的Log数超过100，需要分页返回，总数为{len(qa_infos)}, 请使用page_id参数获取某一页数据，参数范围：[0, {pages - 1}], 本次返回page_id为0的数据"
-    #         qa_infos = qa_infos[:100]
-    #         page_id = 0
-    #     elif page_id >= pages:
-    #         return sanic_json(
-    #             {"code": 2002, "msg": f'输入非法！page_id超过最大值，page_id: {page_id}，最大值：{pages - 1}，请检查！'})
-    #     else:
-    #         msg = f"检索到的Log数超过100，需要分页返回，总数为{len(qa_infos)}, page范围：[0, {pages - 1}], 本次返回page_id为{page_id}的数据"
-    #         qa_infos = qa_infos[page_id * 100:(page_id + 1) * 100]
-    # else:
-    #     msg = f"检索到的Log数为{len(qa_infos)}，一次返回所有数据"
-    #     page_id = 0
-    return sanic_json({"code": 200, "msg": msg, "page_id": page_id, "page_limit": page_limit, "qa_infos": current_qa_infos, "total_count": total_count})
-
-
-@get_time_async
-async def get_random_qa(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    limit = safe_get(req, 'limit', 10)
-    time_start = safe_get(req, 'time_start')
-    time_end = safe_get(req, 'time_end')
-    need_info = safe_get(req, 'need_info')
-    time_range = get_time_range(time_start, time_end)
-    if not time_range:
-        return {"code": 2002, "msg": f'输入非法！time_start格式错误，time_start: {time_start}，示例：2024-10-05，请检查！'}
-
-    debug_logger.info(f"get_random_qa limit: {limit}, time_range: {time_range}")
-    qa_infos = local_doc_qa.milvus_summary.get_random_qa_infos(limit=limit, time_range=time_range, need_info=need_info)
-
-    counts = local_doc_qa.milvus_summary.get_statistic(time_range=time_range)
-    return sanic_json({"code": 200, "msg": "success", "total_users": counts["total_users"],
-                       "total_queries": counts["total_queries"], "qa_infos": qa_infos})
-
-
-@get_time_async
-async def get_related_qa(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    qa_id = safe_get(req, 'qa_id')
-    if not qa_id:
-        return sanic_json({"code": 2005, "msg": "fail, qa_id is None"})
-    need_info = safe_get(req, 'need_info')
-    need_more = safe_get(req, 'need_more', False)
-    debug_logger.info("get_related_qa %s", qa_id)
-    qa_log, recent_logs, older_logs = local_doc_qa.milvus_summary.get_related_qa_infos(qa_id, need_info, need_more)
-    # 按kb_ids划分sections
-    recent_sections = defaultdict(list)
-    for log in recent_logs:
-        recent_sections[log['kb_ids']].append(log)
-    # 把recent_sections的key改为自增的正整数，且每个log都新增kb_name
-    for i, kb_ids in enumerate(list(recent_sections.keys())):
-        kb_names = local_doc_qa.milvus_summary.get_knowledge_base_name(json.loads(kb_ids))
-        kb_names = [kb_name for user_id, kb_id, kb_name in kb_names]
-        kb_names = ','.join(kb_names)
-        recent_sections[i] = recent_sections.pop(kb_ids)
-        for log in recent_sections[i]:
-            log['kb_names'] = kb_names
-
-    older_sections = defaultdict(list)
-    for log in older_logs:
-        older_sections[log['kb_ids']].append(log)
-    # 把older_sections的key改为自增的正整数，且每个log都新增kb_name
-    for i, kb_ids in enumerate(list(older_sections.keys())):
-        kb_names = local_doc_qa.milvus_summary.get_knowledge_base_name(json.loads(kb_ids))
-        kb_names = [kb_name for user_id, kb_id, kb_name in kb_names]
-        kb_names = ','.join(kb_names)
-        older_sections[i] = older_sections.pop(kb_ids)
-        for log in older_sections[i]:
-            log['kb_names'] = kb_names
-
-    return sanic_json({"code": 200, "msg": "success", "qa_info": qa_log, "recent_sections": recent_sections,
-                       "older_sections": older_sections})
-
-
-@get_time_async
-async def get_user_id(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    kb_id = safe_get(req, 'kb_id')
-    kb_id = correct_kb_id(kb_id)
-    debug_logger.info("kb_id: {}".format(kb_id))
-    user_id = local_doc_qa.milvus_summary.get_user_by_kb_id(kb_id)
-    if not user_id:
-        return sanic_json({"code": 2003, "msg": "fail, knowledge Base {} not found".format(kb_id)})
-    else:
-        return sanic_json({"code": 200, "msg": "success", "user_id": user_id})
-
-
-@get_time_async
-async def get_doc(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    doc_id = safe_get(req, 'doc_id')
-    debug_logger.info("get_doc %s", doc_id)
-    if not doc_id:
-        return sanic_json({"code": 2005, "msg": "fail, doc_id is None"})
-    doc_json_data = local_doc_qa.milvus_summary.get_document_by_doc_id(doc_id)
-    return sanic_json({"code": 200, "msg": "success", "doc_text": doc_json_data['kwargs']})
-
-
-@get_time_async
-async def get_rerank_results(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    query = safe_get(req, 'query')
-    if not query:
-        return sanic_json({"code": 2005, "msg": "fail, query is None"})
-    doc_ids = safe_get(req, 'doc_ids')
-    doc_strs = safe_get(req, 'doc_strs')
-    if not doc_ids and not doc_strs:
-        return sanic_json({"code": 2005, "msg": "fail, doc_ids is None and doc_strs is None"})
-    if doc_ids:
-        rerank_results = await local_doc_qa.get_rerank_results(query, doc_ids=doc_ids)
-    else:
-        rerank_results = await local_doc_qa.get_rerank_results(query, doc_strs=doc_strs)
-
-    return sanic_json({"code": 200, "msg": "success", "rerank_results": format_source_documents(rerank_results)})
-
-
-@get_time_async
-async def get_user_status(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("get_user_status %s", user_id)
-    user_status = local_doc_qa.milvus_summary.get_user_status(user_id)
-    if user_status is None:
-        return sanic_json({"code": 2003, "msg": "fail, user {} not found".format(user_id)})
-    if user_status == 0:
-        status = 'green'
-    else:
-        status = 'red'
-    return sanic_json({"code": 200, "msg": "success", "status": status})
-
-
-@get_time_async
-async def health_check(req: request):
-    # 实现一个服务健康检查的逻辑，正常就返回200，不正常就返回500
-    return sanic_json({"code": 200, "msg": "success"})
-
-
-@get_time_async
-async def get_bot_info(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    bot_id = safe_get(req, 'bot_id')
-    if bot_id:
-        if not local_doc_qa.milvus_summary.check_bot_is_exist(bot_id):
-            return sanic_json({"code": 2003, "msg": "fail, Bot {} not found".format(bot_id)})
-    debug_logger.info("get_bot_info %s", user_id)
-    bot_infos = local_doc_qa.milvus_summary.get_bot(user_id, bot_id)
-    data = []
-    for bot_info in bot_infos:
-        if bot_info[6] != "":
-            kb_ids = bot_info[6].split(',')
-            kb_infos = local_doc_qa.milvus_summary.get_knowledge_base_name(kb_ids)
-            kb_names = []
-            for kb_id in kb_ids:
-                for kb_info in kb_infos:
-                    if kb_id == kb_info[1]:
-                        kb_names.append(kb_info[2])
-                        break
-        else:
-            kb_ids = []
-            kb_names = []
-        info = {"bot_id": bot_info[0], "user_id": user_id, "bot_name": bot_info[1], "description": bot_info[2],
-                "head_image": bot_info[3], "prompt_setting": bot_info[4], "welcome_message": bot_info[5],
-                "kb_ids": kb_ids, "kb_names": kb_names,
-                "update_time": bot_info[7].strftime("%Y-%m-%d %H:%M:%S"), "llm_setting": bot_info[9]}
-        data.append(info)
-    return sanic_json({"code": 200, "msg": "success", "data": data})
-
-
-@get_time_async
-async def new_bot(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    bot_name = safe_get(req, "bot_name")
-    desc = safe_get(req, "description", BOT_DESC)
-    head_image = safe_get(req, "head_image", BOT_IMAGE)
-    prompt_setting = safe_get(req, "prompt_setting", BOT_PROMPT)
-    welcome_message = safe_get(req, "welcome_message", BOT_WELCOME)
-    kb_ids = safe_get(req, "kb_ids", [])
-    kb_ids_str = ",".join(kb_ids)
-
-    not_exist_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, kb_ids)
-    if not_exist_kb_ids:
-        msg = "invalid kb_id: {}, please check...".format(not_exist_kb_ids)
-        return sanic_json({"code": 2001, "msg": msg, "data": [{}]})
-    debug_logger.info("new_bot %s", user_id)
-    bot_id = 'BOT' + uuid.uuid4().hex
-    local_doc_qa.milvus_summary.new_qanything_bot(bot_id, user_id, bot_name, desc, head_image, prompt_setting,
-                                                  welcome_message, kb_ids_str)
-    create_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-    return sanic_json({"code": 200, "msg": "success create qanything bot {}".format(bot_id),
-                       "data": {"bot_id": bot_id, "bot_name": bot_name, "create_time": create_time}})
-
-
-@get_time_async
-async def delete_bot(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("delete_bot %s", user_id)
-    bot_id = safe_get(req, 'bot_id')
-    if not local_doc_qa.milvus_summary.check_bot_is_exist(bot_id):
-        return sanic_json({"code": 2003, "msg": "fail, Bot {} not found".format(bot_id)})
-    local_doc_qa.milvus_summary.delete_bot(user_id, bot_id)
-    return sanic_json({"code": 200, "msg": "Bot {} delete success".format(bot_id)})
-
-
-@get_time_async
-async def update_bot(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("update_bot %s", user_id)
-    bot_id = safe_get(req, 'bot_id')
-    if not local_doc_qa.milvus_summary.check_bot_is_exist(bot_id):
-        return sanic_json({"code": 2003, "msg": "fail, Bot {} not found".format(bot_id)})
-    bot_info = local_doc_qa.milvus_summary.get_bot(user_id, bot_id)[0]
-    bot_name = safe_get(req, "bot_name", bot_info[1])
-    description = safe_get(req, "description", bot_info[2])
-    head_image = safe_get(req, "head_image", bot_info[3])
-    prompt_setting = safe_get(req, "prompt_setting", bot_info[4])
-    welcome_message = safe_get(req, "welcome_message", bot_info[5])
-    kb_ids = safe_get(req, "kb_ids")
-    if kb_ids is not None:
-        not_exist_kb_ids = local_doc_qa.milvus_summary.check_kb_exist(user_id, kb_ids)
-        if not_exist_kb_ids:
-            msg = "invalid kb_id: {}, please check...".format(not_exist_kb_ids)
-            return sanic_json({"code": 2001, "msg": msg, "data": [{}]})
-        kb_ids_str = ",".join(kb_ids)
-    else:
-        kb_ids_str = bot_info[6]
-
-    llm_setting = json.loads(bot_info[9])
-    if api_base := safe_get(req, "api_base"):
-        llm_setting["api_base"] = api_base
-    if api_key := safe_get(req, "api_key"):
-        llm_setting["api_key"] = api_key
-    if api_context_length := safe_get(req, "api_context_length"):
-        llm_setting["api_context_length"] = api_context_length
-    if top_p := safe_get(req, "top_p"):
-        llm_setting["top_p"] = top_p
-    if top_k := safe_get(req, "top_k"):
-        llm_setting["top_k"] = top_k
-    if chunk_size := safe_get(req, "chunk_size"):
-        llm_setting["chunk_size"] = chunk_size
-    if temperature := safe_get(req, "temperature"):
-        llm_setting["temperature"] = temperature
-    if model := safe_get(req, "model"):
-        llm_setting["model"] = model
-    if max_token := safe_get(req, "max_token"):
-        llm_setting["max_token"] = max_token
-    # 如果rerank不是None，赋值，false也可以
-    rerank = safe_get(req, "rerank")
-    if rerank is not None:
-        llm_setting["rerank"] = rerank
-    hybrid_search = safe_get(req, "hybrid_search")
-    if hybrid_search is not None:
-        llm_setting["hybrid_search"] = hybrid_search
-    networking = safe_get(req, "networking")
-    if networking is not None:
-        llm_setting["networking"] = networking
-    only_need_search_results = safe_get(req, "only_need_search_results")
-    if only_need_search_results is not None:
-        llm_setting["only_need_search_results"] = only_need_search_results
-
-    debug_logger.info(f"update llm_setting: {llm_setting}")
-
-    # 判断哪些项修改了
-    if bot_name != bot_info[1]:
-        debug_logger.info(f"update bot name from {bot_info[1]} to {bot_name}")
-    if description != bot_info[2]:
-        debug_logger.info(f"update bot description from {bot_info[2]} to {description}")
-    if head_image != bot_info[3]:
-        debug_logger.info(f"update bot head_image from {bot_info[3]} to {head_image}")
-    if prompt_setting != bot_info[4]:
-        debug_logger.info(f"update bot prompt_setting from {bot_info[4]} to {prompt_setting}")
-    if welcome_message != bot_info[5]:
-        debug_logger.info(f"update bot welcome_message from {bot_info[5]} to {welcome_message}")
-    if kb_ids_str != bot_info[6]:
-        debug_logger.info(f"update bot kb_ids from {bot_info[6]} to {kb_ids_str}")
-    #  update_time     TIMESTAMP DEFAULT CURRENT_TIMESTAMP 根据这个mysql的格式获取现在的时间
-    update_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-    debug_logger.info(f"update_time: {update_time}")
-    local_doc_qa.milvus_summary.update_bot(user_id, bot_id, bot_name, description, head_image, prompt_setting,
-                                           welcome_message, kb_ids_str, update_time, llm_setting)
-    return sanic_json({"code": 200, "msg": "Bot {} update success".format(bot_id)})
-
-
-@get_time_async
-async def update_chunks(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    user_id = safe_get(req, 'user_id')
-    user_info = safe_get(req, 'user_info', "1234")
-    passed, msg = check_user_id_and_user_info(user_id, user_info)
-    if not passed:
-        return sanic_json({"code": 2001, "msg": msg})
-    # user_id = user_id + '__' + user_info
-    debug_logger.info("update_chunks %s", user_id)
-    doc_id = safe_get(req, 'doc_id')
-    debug_logger.info(f"doc_id: {doc_id}")
-    yellow_files = local_doc_qa.milvus_summary.get_files_by_status("yellow")
-    if len(yellow_files) > 0:
-        return sanic_json({"code": 2002, "msg": f"fail, currently, there are {len(yellow_files)} files being parsed, please wait for all files to finish parsing before updating the chunk."})
-    update_content = safe_get(req, 'update_content')
-    debug_logger.info(f"update_content: {update_content}")
-    chunk_size = safe_get(req, 'chunk_size', DEFAULT_PARENT_CHUNK_SIZE)
-    debug_logger.info(f"chunk_size: {chunk_size}")
-    update_content_tokens = num_tokens_embed(update_content)
-    if update_content_tokens > chunk_size:
-        return sanic_json({"code": 2003, "msg": f"fail, update_content too long, please reduce the length, "
-                                                f"your update_content tokens is {update_content_tokens}, "
-                                                f"the max tokens is {chunk_size}"})
-    doc_json = local_doc_qa.milvus_summary.get_document_by_doc_id(doc_id)
-    if not doc_json:
-        return sanic_json({"code": 2004, "msg": "fail, DocId {} not found".format(doc_id)})
-    doc = Document(page_content=update_content, metadata=doc_json['kwargs']['metadata'])
-    doc.metadata['doc_id'] = doc_id
-    local_doc_qa.milvus_summary.update_document(doc_id, update_content)
-    expr = f'doc_id == "{doc_id}"'
-    local_doc_qa.milvus_kb.delete_expr(expr)
-    await local_doc_qa.retriever.insert_documents([doc], chunk_size, True)
-    return sanic_json({"code": 200, "msg": "success update doc_id {}".format(doc_id)})
-
-
-@get_time_async
-async def get_file_base64(req: request):
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    file_id = safe_get(req, 'file_id')
-    debug_logger.info("get_file_base64 %s", file_id)
-    file_location = local_doc_qa.milvus_summary.get_file_location(file_id)
-    debug_logger.info("file_location %s", file_location)
-    # file_location = '/home/liujx/Downloads/2021-08-01 00:00:00.pdf'
-    if not file_location:
-        return sanic_json({"code": 2005, "msg": "fail, file_id is Invalid"})
-    with open(file_location, "rb") as f:
-        file_base64 = base64.b64encode(f.read()).decode()
-    return sanic_json({"code": 200, "msg": "success", "file_base64": file_base64})
-
-# 获取用户角色
-def get_user_role(user_id, req: request):
-    # 从数据库获取用户角色
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    query = "SELECT role FROM User WHERE user_id = %s"
-    result = local_doc_qa.milvus_summary.execute_query_(query, (user_id,), fetch=True)
-    if result and result[0]:
-        role = result[0][0]
-        debug_logger.info(f"获取用户角色 - 用户: {user_id}, 角色: {role}")
-        return role
-    debug_logger.warning(f"未找到用户角色，使用默认角色 'user' - 用户: {user_id}")
-    return "user"  # 默认角色
-
-# 检查用户是否有权限
-def has_permission(user_role, required_permission):
-    # 权限等级
-    permission_levels = {
-        "read": 1,
-        "write": 2,
-        "admin": 3,
-    }
-    
-    # 角色对应的权限等级
-    role_permissions = {
-        "user": 1,        # 普通用户只有读权限
-        "editor": 2,      # 编辑者有写权限
-        "admin": 3,       # 管理员有管理权限
-    }
-    
-    user_level = role_permissions.get(user_role, 0)
-    required_level = permission_levels.get(required_permission, 0)
-    has_perm = user_level >= required_level
-    
-    debug_logger.info(f"权限检查 - 用户角色: {user_role}({user_level}), 所需权限: {required_permission}({required_level}), 结果: {'通过' if has_perm else '不通过'}")
-    return has_perm
-
-@get_time_async
-async def login(req: request):
-    """用户登录接口，验证用户身份并生成JWT令牌"""
-    local_doc_qa: LocalDocQA = req.app.ctx.local_doc_qa
-    email = safe_get(req, 'email')
-    password = safe_get(req, 'password')
-    
-    debug_logger.info(f"用户登录 - 邮箱: {email}")
-    
-    if not email or not password:
-        debug_logger.error("登录失败 - 邮箱和密码不能为空")
-        return response.json({"code": 400, "msg": "邮箱和密码不能为空"})
-    
-    # 验证用户
-    query = "SELECT user_id, password, role, user_name, dept_id FROM User WHERE email = %s AND status = 'active'"
-    result = local_doc_qa.milvus_summary.execute_query_(query, (email,), fetch=True)
-    
-    if not result:
-        debug_logger.error(f"登录失败 - 用户不存在或未激活: {email}")
-        return response.json({"code": 401, "msg": "用户不存在或未激活"})
-    
-    user_id, stored_password, role, user_name, dept_id = result[0]
-    debug_logger.info(f"找到用户 - ID: {user_id}, 角色: {role}")
-    
-    # 验证密码
-    try:
-        # 如果存储的密码是明文（初始管理员密码）
-        if stored_password == password:
-            debug_logger.info("使用明文密码验证成功，正在升级为加密密码")
-            # 升级为加密密码
-            salt = bcrypt.gensalt()
-            hashed_password = bcrypt.hashpw(password.encode('utf-8'), salt)
-            update_query = "UPDATE User SET password = %s WHERE user_id = %s"
-            local_doc_qa.milvus_summary.execute_query_(update_query, (hashed_password.decode('utf-8'), user_id), commit=True)
-            password_valid = True
-        else:
-            # 尝试验证加密密码
-            try:
-                password_valid = bcrypt.checkpw(password.encode('utf-8'), stored_password.encode('utf-8'))
-            except Exception as e:
-                debug_logger.error(f"密码验证出错（可能是密码格式问题）: {str(e)}")
-                password_valid = False
-        
-        if not password_valid:
-            debug_logger.error(f"登录失败 - 密码错误: {email}")
-            return response.json({"code": 401, "msg": "密码错误"})
-            
-    except Exception as e:
-        debug_logger.error(f"密码验证过程出错: {str(e)}")
-        return response.json({"code": 500, "msg": "密码验证过程出错"})
-    
-    # 获取部门信息
-    dept_name = None
-    if dept_id:
-        dept_query = "SELECT dept_name FROM Department WHERE dept_id = %s"
-        dept_result = local_doc_qa.milvus_summary.execute_query_(dept_query, (dept_id,), fetch=True)
-        if dept_result:
-            dept_name = dept_result[0][0]
-    
-    # 生成JWT令牌
-    payload = {
-        "user_id": user_id,
-        "role": role,
-        "dept_id": dept_id,
-        "exp": datetime.utcnow() + JWT_ACCESS_TOKEN_EXPIRES
-    }
-    token = jwt.encode(payload, JWT_SECRET_KEY, algorithm=JWT_ALGORITHM)
-    
-    debug_logger.info(f"登录成功 - 用户: {user_id}, 角色: {role}, 部门: {dept_name}")
-    return response.json({
-        "code": 200,
-        "msg": "登录成功",
-        "data": {
-            "token": token,
-            "user_id": user_id,
-            "user_name": user_name,
-            "role": role,
-            "dept_id": dept_id,
-            "dept_name": dept_name
-        }
-    })
-
-def verify_password(plain_password, hashed_password):
-    """验证密码是否匹配"""
-    try:
-        # 使用 bcrypt 进行密码验证
-        result = bcrypt.checkpw(plain_password.encode(), hashed_password.encode())
-        debug_logger.info(f"密码验证{'成功' if result else '失败'}")
-        return result
-    except Exception as e:
-        debug_logger.error(f"密码验证出错: {str(e)}")
-        return False
-
-@get_time_async
-@auth_required("read")  # 刷新令牌只需要基本权限
-async def refresh_token(req: request):
-    """刷新JWT令牌"""
-    user_id = req.ctx.user["user_id"]
-    role = req.ctx.user["role"]
-    
-    debug_logger.info(f"刷新令牌 - 用户: {user_id}, 角色: {role}")
-    
-    # 生成新的JWT令牌
-    payload = {
-        "user_id": user_id,
-        "role": role,
-        "exp": datetime.utcnow() + JWT_ACCESS_TOKEN_EXPIRES
-    }
-    token = jwt.encode(payload, JWT_SECRET_KEY, algorithm=JWT_ALGORITHM)
-    
-    debug_logger.info(f"令牌刷新成功 - 用户: {user_id}")
-    return response.json({
-        "code": 200,
-        "msg": "令牌刷新成功",
-        "data": {
-            "token": token
-        }
-    })
-
